+++ .github/workflows/ci.yml
@@ -48,7 +48,6 @@ jobs:
         with:
           version: v1.51.0
 
-
   unit-tests:
     name: Test packages
     runs-on: ubuntu-latest
@@ -62,9 +61,24 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Test
-        run: make test-with-cover
+        run: make test-unit
 
   integration-tests:
+    name: Test packages
+    runs-on: ubuntu-latest
+    steps:
+      - name: Set up Go 1.20
+        uses: actions/setup-go@v3
+        with:
+          go-version: 1.20.x
+
+      - name: Check out code
+        uses: actions/checkout@v3
+
+      - name: Test
+        run: make test-integration
+
+  e2e-tests:
     name: Test integration e2e suite
     runs-on: ubuntu-latest
     steps:
@@ -79,7 +93,7 @@ jobs:
       - name: Test
         run: make test-e2e
 
-  integration-tests-serverless:
+  e2e-tests-serverless:
     name: Test serverless integration e2e suite
     runs-on: ubuntu-latest
     steps:
